backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-environ==0.11.2
django-redis==5.4.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
Pillow==10.1.0
boto3==1.34.11
sentry-sdk==1.39.1
tensorflow==2.15.0
numpy==1.26.2
python-dotenv==1.0.0
twilio==8.10.3
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.12.1
flake8==6.1.0
isort==5.13.2
